==> Notebooks/1_Extraccion_webscrap.ipynb <==
## EXTRACCIÓN DE LOS DATOS:



En primer lugar, comenzamos realizando **web scraping** de la página web del Ministerio de Igualdad donde se encuentran hipervinculadas las campañas 
realizadas sobre prevención de la violencia de género realizadas en los años 2020 a 2023. El objetivo es conseguir la **fecha de publicación** de cada campaña, redirigiéndonos desde esta url a las páginas de YouTube donde se encuentran publicados los vídeos.

Fuente: [Delegación del Gobierno contra la Violencia de Género](https://violenciagenero.igualdad.gob.es/sensibilizacionConcienciacion/campannas/violenciaGobierno/home.htm)

Durante la investigación se descubre un artículo que asegura que la publicación por televisión pública de un documental sobre un testimonio de violencia de género tuvo impacto en las llamadas recibidas por el 016, por lo que se decide contarlo como una campaña más. Para ello, se realiza scrapeo desde el artículo donde se publica esta noticia para llegar a la página donde viene registrada la **fecha de emisión** del documental.

Fuente: [La mayor subida intermensual de llamadas al 016 de la historia: subieron un 61% en marzo coincidiendo con la emisión del documental de Rocío Carrasco](https://maldita.es/malditodato/20210409/llamadas-016-crecieron-marzo-documental-rocio-carrasco/)


import pandas as pd
from selenium import webdriver
from selenium.webdriver.chrome.options import Options
from webdriver_manager.chrome import ChromeDriverManager
from selenium.webdriver.support.ui import WebDriverWait
from selenium.common.exceptions import NoSuchElementException
import time
from datetime import datetime
import re

opciones=Options()

opciones.add_experimental_option('excludeSwitches', ['enable-automation'])
opciones.add_experimental_option('useAutomationExtension', False)
opciones.headless=False   
opciones.add_argument('--start-maximized')         
opciones.add_argument('--incognito')


from selenium import webdriver
PATH = "/driver/chromedriver"
from selenium.webdriver.common.by import By
driver = webdriver.Chrome(opciones)

### 1.   WEB SCRAPING *CAMPAÑAS DE PREVENCIÓN CONTRA LA VIOLENCIA DE GÉNERO (Ministerio de Igualdad)*

# Accedemos a la url:

url = 'https://violenciagenero.igualdad.gob.es/sensibilizacionConcienciacion/campannas/violenciaGobierno/home.htm'

driver.get(url)

# Accedemos a las urls que redirigen a las campañas publicadas en YouTube:

urls = driver.find_elements(By.XPATH, '//*[@id="contenido"]/div[3]/div/ul/li/a')

urlss = [i.get_attribute('href') for i in urls]

urlss = urlss[:9]

urlss

# Creamos un diccionario para almacenar las urls de cada campaña:

campañas = {}

for i in range(len(urlss)):
    campañas[f"campaña{i + 1}"] = urlss[i]
campañas

driver.quit()

def extraer_fecha(url):
    driver = webdriver.Chrome()
    driver.get(url)
    time.sleep(3)
    driver.find_element(By.XPATH, '//*[@id="mensaje"]/div[1]/a').click() # Clickamos sobre el mensaje de inicio para entrar en la web
    time.sleep(4)
    frame = driver.find_element(By.CSS_SELECTOR, 'iframe.player') # El siguiente paso es darle al play en la miniatura
    driver.switch_to.frame(frame)    # Para ello hay que entrar con el driver en el propio frame del vídeo
    time.sleep(3)
    driver.find_element(By.XPATH, '//*[@id="movie_player"]/div[4]/button').click() # Ahora sí le damos al play
    driver.find_element(By.CSS_SELECTOR, '#movie_player > div.ytp-chrome-bottom > div.ytp-chrome-controls > div.ytp-right-controls > a').click() # Nos redirige a Youtube
    driver.switch_to.window(driver.window_handles[1])  # Para cambiar de ventana, ya que hemos abierto una nueva
    time.sleep(3)
    driver.find_element(By.XPATH, '//*[@id="content"]/div[2]/div[6]/div[1]/ytd-button-renderer[2]/yt-button-shape').click()  # Aceptamos cookies
    time.sleep(2)
    driver.find_element(By.CSS_SELECTOR, '#movie_player > div.ytp-chrome-bottom > div.ytp-chrome-controls > div.ytp-left-controls > button').click()   # Pausamos el vídeo
    time.sleep(3)
    driver.find_element(By.XPATH, '//*[@id="expand"]').click()    # Desplegamos para ver la fecha
    fecha = [driver.find_element(By.XPATH, '//*[@id="info"]/span[3]').text]   # Extraemos la fecha
    driver.quit()
    return fecha

# Aplicamos la función creada 'extraer_fechas' para aplicarlo en bucle a todas las urls de las campañas:

fechas = []
for campaña, url in campañas.items():
    fechas.append(extraer_fecha(url))

# Comprobamos las fechas obtenidas:

fechas

# Cambiamos el tipo de dato a datetime:

fecha_campaña = [datetime.strptime(fecha[0], '%d %b %Y').date() for fecha in fechas]

fecha_campaña

# Creamos el dataframe con el año, el mes y las urls de las campañas:

campañas_vg = pd.DataFrame({'Año': [i.year for i in fecha_campaña], 
                            'Mes': [i.month for i in fecha_campaña], 
                            'URL': campañas.values()})
campañas_vg

# Visualizamos la información del dataframe final:

campañas_vg.info()

### 2. WEB SCRAPING DOCUMENTAL *Rocío, contar la verdad para seguir viva*:

# Accedemos a la url para scrapear la fecha de emisión:

url_ori = 'https://maldita.es/malditodato/20210409/llamadas-016-crecieron-marzo-documental-rocio-carrasco/'

driver = webdriver.Chrome()
driver.get(url_ori)
time.sleep(3)

driver.find_element(By.XPATH, '//*[@id="iuxh"]/div[1]/div/div[2]/button[1]').click()   # Aceptamos cookies
time.sleep(2)

driver.find_element(By.XPATH, '//*[@id="article-text"]/p[1]/a[2]').click()  #Pulsamos sobre el hipervínculo 'emisión del primer episodio el 21 de marzo'
time.sleep(3)
driver.switch_to.window(driver.window_handles[1])  # Cambiamos de ventana, ya que hemos abierto una nueva

driver.find_element(By.XPATH, '//*[@id="didomi-notice-agree-button"]').click()  # Aceptamos cookies de nuevo
time.sleep(2)

titular = driver.find_element(By.XPATH, '//*[@id="article_multisite"]/div/article/header/h1').text  # Clickamos sobre el texto donde se refleja la fecha

fecha_docu = titular.split()[-1]   # Dividimos la string por espacios y nos quedamos con el último elemento, que es la fecha
driver.quit()

# Visualizamos la fecha extraída:

fecha_docu

fecha_docu = fecha_docu.split('/')   # Dividimos para poder obtener por separado el mes y el año
fecha_docu

mes_docu = int(fecha_docu[1])     # Obtenemos el mes
mes_docu

año_docu = re.search(r'\d+', fecha_docu[2]).group()   # Nos quedamos solo con la parte numérica del año
año_docu

año_doc = int(datetime.strptime(año_docu, "%y").strftime("%Y"))   # Lo convertimos de tipo para que tenga consistencia con los anteriores
año_doc

# Creamos una nueva columna para añadirla al dataframe de las campañas:

nueva_fila = pd.DataFrame({'Año': [año_doc], 'Mes': [mes_docu], 'URL': ['https://www.mitele.es/programas-tv/rocio-contar-la-verdad-para-seguir-viva/']})

campañas_vg = pd.concat([campañas_vg, nueva_fila], ignore_index=True)
campañas_vg.info()

### 3. DATAFRAME FINAL *campañas_vg*:

campañas_vg = campañas_vg.sort_values(by='Año')
campañas_vg

# Exportamos el archivo:

campañas_vg.to_csv('../data/data_clean/campañas_vg.csv', index=False)
