==> Notebooks/2_Transformacion_datos.ipynb <==
## **TRANSFORMACIÓN Y LIMPIEZA DE LOS DATOS:**

Continuamos el proyecto con la descarga de tres archivos en formato csv para proceder a su transformación y limpieza, relativos a:
- Número de llamadas recibidas por el 016
- Número de denuncias interpuestas.
- Número de víctimas mortales. 

No existen nulos ni duplicados, por lo que la limpieza se basa en transformar el tipo de dato de las columnas Año y Mes, cambiar el tipo en el número de llamadas, denuncias y víctimas a entero y cambiar el nombre de estas tres columnas en los tres archivos.
También eliminamos aquellos datos anteriores a 2020, para ceñirnos a aquellos años en que hubo campañas de prevención (entre 2020 y 2023) y creamos la columna de Trimestre para los casos en que no existe.

import pandas as pd
import warnings
warnings.filterwarnings('ignore')
pd.set_option('display.max_columns', None)

### **1. REGISTRO DE LLAMADAS RECIBIDAS POR EL 016:**

El Ministerio de Igualdad, por medio de la Delegación del Gobierno contra la Violencia de Género, presta el servicio telefónico de información, de asesoramiento y apoyo psicosocial para la atención de todas las formas de violencia contra las mujeres.

Tras la descarga de un archivo csv que contiene el registro de las llamadas recibidas según provincia, mes y año, se limpia el documento para obtener únicamente las llamadas recibidas entre 2020 y 2023, ya que estos son los años en que fueron publicadas las campañas de prevención.

Fuente: [Portal Estadístico Delegación del Gobierno contra la Violencia de Género](https://estadisticasviolenciagenero.igualdad.gob.es/)

llamadas = pd.read_csv('/Users/noeliarosonmartin/Ironhack/Project_ETL/data/data_original/llamadas.csv')
llamadas.head()

llamadas['Año'] = llamadas['Año'].str.replace(r'\D', '', regex=True) # Eliminamos lo que no es numérico para que tenga consistencia con las otras tablas

llamadas['Año'] = pd.to_datetime(llamadas['Año'], format='%Y').dt.year

llamadas = llamadas.loc[llamadas['Año'] >= 2020]   # Eliminamos los años anteriores a 2020
llamadas['Año'].unique()

llamadas['Mes'].unique()

# Transformamos los meses a formato numérico para mantener la consistencia:

meses_año = {
    'Enero': '01',
    'Febrero': '02',
    'Marzo': '03',
    'Abril': '04',
    'Mayo': '05',
    'Junio': '06',
    'Julio': '07',
    'Agosto': '08',
    'Septiembre': '09',
    'Octubre': '10',
    'Noviembre': '11',
    'Diciembre': '12'}

llamadas['Mes'] = llamadas['Mes'].map(meses_año)

llamadas['Mes'] = llamadas['Mes'].astype(int)

llamadas.rename(columns={'Número de llamadas pertinentes': 'Numero_llamadas'}, inplace=True)
llamadas['Numero_llamadas'] = llamadas['Numero_llamadas'].astype(int)

# Función creada para crear una nueva columna en la que aparezca el trimestre del año al que pertenece.
# Entra un mes y sale el número del trimestre

def trimestre(mes):
    if mes in [1, 2, 3]:
        return 1
    elif mes in [4, 5, 6]:
        return 2
    elif mes in [7, 8, 9]:
        return 3
    else:
        return 4

# Aplicamos la función trimestre:

llamadas['Trimestre'] = llamadas['Mes'].apply(trimestre)

# Reordenamos las columnas:

llamadas = llamadas [['Provincia', 'Año', 'Trimestre', 'Mes', 'Numero_llamadas']]

llamadas.duplicated().sum()

llamadas.isna().sum()

llamadas.info(memory_usage = 'deep')

### **DATAFRAME FINAL *llamadas.csv***

llamadas = llamadas.reset_index(drop=True)
llamadas.head()

# Exportamos csv:

llamadas.to_csv('../data/data_clean/llamadas_016.csv', index=False)

### **2. REGISTRO DE DENUNCIAS POR VIOLENCIA DE GÉNERO:**

El Ministerio de Igualdad publica el registro de denuncias en base a varios criterios, seleccionando en este caso para nuestro proyecto el número de denuncias según año, mes y provincia. Al igual que con la tabla anterior, se dejan los datos relativos a los años 2020 a 2023 y se limpian los datos de forma que mantenga la consistencia anterior.

Fuente: [Portal Estadístico Delegación del Gobierno contra la Violencia de Género](https://estadisticasviolenciagenero.igualdad.gob.es/)

denuncias = pd.read_csv('/Users/noeliarosonmartin/Ironhack/Project_ETL/data/data_original/denunc.csv')
denuncias

denuncias.info(memory_usage='deep')

denuncias['Año'] = denuncias['Año'].str.replace(r'\D', '', regex=True) # Eliminamos lo que no es numérico

denuncias['Año'] = pd.to_datetime(denuncias['Año'], format='%Y').dt.year  # Cambiamos a tipo datetime

denuncias = denuncias.loc[denuncias['Año'] >= 2020]  # Eliminamos los años anteriores a 2020

# Modificamos el formato de la columna Trimestre:

trimestr = {
    'Primero': 1,
    'Segundo': 2,
    'Tercero': 3,
    'Cuarto': 4}

denuncias['Trimestre'] = denuncias['Trimestre'].map(trimestr)


denuncias.rename(columns={'Número de denuncias por violencia de género': 'Numero_denuncias'}, inplace=True) # Renombramos la columna
denuncias['Numero_denuncias'] = denuncias['Numero_denuncias'].astype(int)  # Cambiamos a tipo entero


denuncias.duplicated().sum()

denuncias.isna().sum()

denuncias.info(memory_usage='deep')

### **DATAFRAME FINAL *denuncias.csv***

denuncias = denuncias.reset_index(drop=True)
denuncias.head()

# EXPORTAR CSV
denuncias.to_csv('../data/data_clean/denuncias_vg.csv', index=False)

### **3. REGISTRO DE VÍCTIMAS MORTALES POR VIOLENCIA DE GÉNERO:**

El Ministerio de Igualdad publica el registro de víctimas mortales en base a varios criterios, seleccionando en este caso para nuestro proyecto el número de víctimas según año, mes y provincia. Al igual que con la tabla anterior, se dejan los datos relativos a los años 2020 a 2023 y se limpian los datos de forma que mantenga la consistencia anterior.

Fuente: [Portal Estadístico Delegación del Gobierno contra la Violencia de Género](https://estadisticasviolenciagenero.igualdad.gob.es/)

victimas = pd.read_csv('/Users/noeliarosonmartin/Ironhack/Project_ETL/data/data_original/victimas_mort.csv')
victimas

victimas.info(memory_usage='deep')

victimas['Año'] = victimas['Año'].str.replace(r'\D', '', regex=True) # Eliminamos lo que no es numérico

victimas['Año'] = pd.to_datetime(victimas['Año'], format='%Y').dt.year # Cambiamos a tipo datetime

victimas = victimas.loc[victimas['Año'] >= 2020]   # Eliminamos los años anteriores a 2020

victimas['Mes'] = victimas['Mes'].map(meses_año)   # Utilizamos el diccionario anterior para aplicar formato numérico a los meses

victimas['Mes'] = victimas['Mes'].astype(int) 

victimas['Trimestre'] = victimas['Mes'].apply(trimestre)  # Utilizamos la función anterior para crear la nueva columna de Trimestre


# Renombramos columna de número de víctimas y pasamos a tipo entero 

victimas.rename(columns={'Número de mujeres víctimas mortales': 'Num_victimas_mortales'}, inplace=True)
victimas['Num_victimas_mortales'] = victimas['Num_victimas_mortales'].astype(int)

# Reordenamos las columnas

victimas = victimas [['Provincia', 'Año', 'Trimestre', 'Mes', 'Num_victimas_mortales']]

victimas.isna().sum()

victimas.duplicated().sum()

victimas.info(memory_usage='deep')

### **DATAFRAME FINAL *victimas.csv***

victimas = victimas.reset_index(drop=True)
victimas.head()

# EXPORTAR CSV
victimas.to_csv('../data/data_clean/victimas_mortales_vg.csv', index=False)
