==> Notebooks/3_Carga_MongoDB.ipynb <==
### **CARGA EN BASE DE DATOS: *MONGODB***

MongoDB es un sistema de base de datos NoSQL, orientado a documentos y de código abierto que, en lugar de guardar los datos en tablas, los guarda en estructuras de datos BSON, similar a JSON.

Se cargan en este sistema los cuatro archivos csv generados para futuros análisis de los mismos.


import pandas as pd
from pymongo import MongoClient


llam = pd.read_csv('/Users/noeliarosonmartin/Ironhack/Project_ETL/data/data_clean/llamadas_016.csv')
denu = pd.read_csv('/Users/noeliarosonmartin/Ironhack/Project_ETL/data/data_clean/denuncias_vg.csv')
vict = pd.read_csv('/Users/noeliarosonmartin/Ironhack/Project_ETL/data/data_clean/victimas_mortales_vg.csv')
camp = pd.read_csv('/Users/noeliarosonmartin/Ironhack/Project_ETL/data/data_clean/campañas_vg.csv')


client = MongoClient("mongodb://localhost:27017")
client

# Creación de la base de datos:

db = client['impacto_vg']    

# Inserción de los datos:

db['llamadas'].insert_many(llam.to_dict(orient='records'))
db['denuncias'].insert_many(denu.to_dict(orient='records'))
db['victimas'].insert_many(vict.to_dict(orient='records'))
db['campañas'].insert_many(camp.to_dict(orient='records'))

db.list_collection_names()

#### **¿EN QUÉ MES DE 2020 HUBO CAMPAÑA DE PREVENCIÓN?**


query = { 'Año': 2020}

select = {  'Año': 1,
            '_id': 0,
            'Mes': 1}

resultado = list(db['campañas'].find(query, select))

f'Hubo campaña en {resultado}'



#### **¿QUÉ PROVINCIAS RECIBIERON MÁS LLAMADAS AL 016 EN LOS MESES SIGUIENTES A LA CAMPAÑA DE NOVIEMBRE 2020?**

query = {'Año': 2020,
         'Mes': {'$in': [11, 12]}}     
select = {  'Provincia': 1,
            '_id': 0,
            'Mes': 1,
            'Numero_llamadas': 1}

resultado = list(db['llamadas'].find(query, select).sort('Numero_llamadas', -1).limit(5))

resultado

#### **¿Y SI LO COMPARAMOS CON LOS MESES ANTERIORES?**

Ejemplo: Sevilla

query = {'Año': 2020,
         'Mes': {'$in': [8, 9, 10, 11, 12]},
         'Provincia': 'Sevilla'}     
select = { '_id': 0,
            'Mes': 1,
            'Año': 1,
            'Numero_llamadas': 1}

resultado = list(db['llamadas'].find(query, select).limit(10))

resultado

Ejemplo: Barcelona

query = {'Año': 2020,
         'Mes': {'$in': [8, 9, 10, 11, 12]},
         'Provincia': 'Barcelona'}     
select = { '_id': 0,
            'Mes': 1,
            'Año': 1,
            'Numero_llamadas': 1}

resultado = list(db['llamadas'].find(query, select).limit(10))

resultado

Como observamos, en esas dos ciudades las llamadas bajaron tras la primera campaña. Vamos a observar esos dos ejemplos y vamos a comprobar las llamadas en esos meses del año siguiente. Pero primero comprobaremos cuántas campañas hubo durante el 2021.

query = { 'Año': 2021}

select = {'_id': 0,
          'Mes': 1}

resultado = list(db['campañas'].find(query, select))
Mes = resultado

f'Hubo {len(resultado)} campañas en los meses {Mes} de 2021'

query = {'Año': 2021,
         'Mes': {'$in': [8, 9, 10, 11, 12]},
         'Provincia': 'Sevilla'}     
select = { '_id': 0,
            'Mes': 1,
            'Numero_llamadas': 1}

resultado = list(db['llamadas'].find(query, select).limit(10))

resultado

Se observa una pequeña subida en las llamadas en los meses de octubre, noviembre y diciembre de 2021, con respecto a 2020, en Sevilla.

query = {'Año': 2021,
         'Mes': {'$in': [8, 9, 10, 11, 12]},
         'Provincia': 'Barcelona'}     
select = { '_id': 0,
            'Mes': 1,
            'Numero_llamadas': 1}

resultado = list(db['llamadas'].find(query, select).limit(10))

resultado

En Barcelona se observa un incremento generalizado en todos esos meses con respecto al año anterior, si bien es cierto que en los meses de noviembre y diciembre el aumento parece mayor.

Centrándonos en la ciudad de Sevilla, vamos a observar cuál es el mes con más llamadas en todos los años:

query = {'Provincia': 'Sevilla'}     
select = { '_id': 0,
            'Mes': 1,
            'Año': 1,
            'Numero_llamadas': 1}

resultado = list(db['llamadas'].find(query, select).sort('Numero_llamadas', -1).limit(3))

resultado

Y ahora, el mes con menos llamadas en todos los años:

query = {'Provincia': 'Sevilla'}     
select = { '_id': 0,
            'Mes': 1,
            'Año': 1,
            'Numero_llamadas': 1}

resultado = list(db['llamadas'].find(query, select).sort('Numero_llamadas').limit(3))

resultado

#### **¿CUÁL FUE EL TRIMESTRE CON MÁS DENUNCIAS EN LA CIUDAD DE SEVILLA?**

query = {'Provincia': 'Sevilla'}     
select = { '_id': 0,
            'Trimestre': 1,
            'Año': 1,
            'Numero_denuncias': 1}

resultado = list(db['denuncias'].find(query, select).sort('Numero_denuncias', -1).limit(1))

resultado
